requirements.txt
numpy==1.26.4
pandas==2.1.4
pillow==10.2.0
pyarrow==14.0.2
scikit-learn==1.3.2
tensorboard==2.14.0
timm==0.9.12
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
matplotlib==3.8.2
